tests: Move ObjectEventWatcher to a separate package
Moved ObjectEventWatcher from utils.go to a separate package.
