+++ tests/migration_test.go
@@ -85,6 +85,7 @@ import (
 	"kubevirt.io/kubevirt/tests/flags"
 	"kubevirt.io/kubevirt/tests/libnet"
 	"kubevirt.io/kubevirt/tests/libstorage"
+	"kubevirt.io/kubevirt/tests/watcher"
 )
 
 const (
@@ -1187,11 +1188,10 @@ var _ = Describe("[rfe_id:393][crit:high][vendor:cnv-qe@redhat.com][level:system
 				By("Should receive warning event that target pod is currently unschedulable")
 				ctx, cancel := context.WithCancel(context.Background())
 				defer cancel()
-				tests.
-					NewObjectEventWatcher(migration).
+				watcher.New(migration).
 					Timeout(60*time.Second).
 					SinceWatchedObjectResourceVersion().
-					WaitFor(ctx, tests.WarningEvent, "migrationTargetPodUnschedulable")
+					WaitFor(ctx, watcher.WarningEvent, "migrationTargetPodUnschedulable")
 
 				By("Migration should observe a timeout period before canceling unschedulable target pod")
 				Consistently(func() error {
