+++ tests/storage/BUILD.bazel
@@ -43,6 +43,7 @@ go_library(
         "//tests/libvmi:go_default_library",
         "//tests/testsuite:go_default_library",
         "//tests/util:go_default_library",
+        "//tests/watcher:go_default_library",
         "//vendor/github.com/google/goexpect:go_default_library",
         "//vendor/github.com/kubernetes-csi/external-snapshotter/client/v4/apis/volumesnapshot/v1:go_default_library",
         "//vendor/github.com/onsi/ginkgo/v2:go_default_library",
